+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## unreleased
+- Add support for `ectmp` file paths [#2](https://github.com/observingClouds/ecmwfspec/issues/2)
+
 ## 0.0.1
 - Initial release